chore(ui): minor tweaks
